+++ product_portfolio/tests/test_api.py
@@ -9,6 +9,7 @@
 import json
 
 from django.core import mail
+from django.core.files.base import ContentFile
 from django.test import TestCase
 from django.urls import reverse
 
@@ -41,6 +42,7 @@ from product_portfolio.models import ProductItemPurpose
 from product_portfolio.models import ProductPackage
 from product_portfolio.models import ProductRelationStatus
 from product_portfolio.models import ProductStatus
+from product_portfolio.models import ScanCodeProject
 
 
 class ProductAPITestCase(MaxQueryMixin, TestCase):
@@ -1125,3 +1127,30 @@ class ProductRelatedAPITestCase(TestCase):
         from dje.api_custom import TabPermission  # Prevent circular import
 
         self.assertEqual((TabPermission,), CodebaseResourceViewSet.extra_permissions)
+
+    def test_api_product_endpoint_load_sboms_action(self):
+        url = reverse("api_v2:product-load-sboms", args=[self.product1.uuid])
+
+        self.client.login(username=self.base_user.username, password="secret")
+        response = self.client.post(url, data={})
+        self.assertEqual(status.HTTP_403_FORBIDDEN, response.status_code)
+
+        # Required permissions
+        add_perm(self.base_user, "add_product")
+        assign_perm("view_product", self.base_user, self.product1)
+
+        response = self.client.post(url, data={})
+        self.assertEqual(status.HTTP_400_BAD_REQUEST, response.status_code)
+        expected = {"input_file": ["This field is required."]}
+        self.assertEqual(expected, response.data)
+
+        data = {
+            "input_file": ContentFile("Content", name="sbom.json"),
+            "update_existing_packages": False,
+            "scan_all_packages": False,
+        }
+        response = self.client.post(url, data)
+        self.assertEqual(status.HTTP_200_OK, response.status_code)
+        expected = {"status": "SBOM file submitted to ScanCode.io for inspection."}
+        self.assertEqual(expected, response.data)
+        self.assertEqual(1, ScanCodeProject.objects.count())